Fixed bug in interceptor with next()
Turns out that there is a difference between calling next()
in a route and calling return next(). In this case, we wanted
to call return next() so that the rest of the route doesn't run
The alternate approach is to use some sort of if/else construct
to ensure that if you call next(), the rest of the route is in
the else and doesn't execute.